+++ src/main/java/org/binclassreader/attributes/LocalVariableTypeTableAttr.java
@@ -31,7 +31,7 @@ import java.util.List;
 public class LocalVariableTypeTableAttr extends AbstractAttribute implements SelfReader {
 
     @BinClassParser(readOrder = 3, byteToRead = 4)
-    private int[] local_variable_type_table_length;
+    private short[] local_variable_type_table_length;
     private List<LocalVariableType> localVarTable;
 
     public int getLocalVariableTypeLength() {
@@ -54,19 +54,19 @@ public class LocalVariableTypeTableAttr extends AbstractAttribute implements Sel
 
     private class LocalVariableType {
         @BinClassParser(byteToRead = 2)
-        private int[] start_pc;
+        private short[] start_pc;
 
         @BinClassParser(readOrder = 2, byteToRead = 2)
-        private int[] length;
+        private short[] length;
 
         @BinClassParser(readOrder = 3, byteToRead = 2)
-        private int[] name_index;
+        private short[] name_index;
 
         @BinClassParser(readOrder = 4, byteToRead = 2)
-        private int[] signature_index;
+        private short[] signature_index;
 
         @BinClassParser(readOrder = 5, byteToRead = 2)
-        private int[] index;
+        private short[] index;
 
 
         public int getStartPc() {
